smart_store_backend/pyproject.toml

[tool.poetry]
name = "smart_store_backend"
version = "0.1.0"
description = "Backend for the Smart Store application"
authors = ["AI Engineer <ai@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^22.0.0"
google-cloud-firestore = "^2.16.0"
google-cloud-storage = "^2.16.0"
google-cloud-pubsub = "^2.20.0"
google-cloud-secret-manager = "^2.19.0"
google-cloud-aiplatform = "^1.54.0"
firebase-admin = "^6.5.0"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
